@@ -40,16 +43,16 @@
   },
   "homepage": "https://github.com/angular/angularfire#readme",
   "dependencies": {
-    "@angular-devkit/architect": ">= 0.900 < 0.1300",
-    "@angular-devkit/core": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "@angular-devkit/schematics": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "@angular/common": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "@angular/compiler": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "@angular/core": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "@angular/platform-browser": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "@angular/platform-browser-dynamic": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "@angular/router": "^9.0.0 || ^10.0.0 || ^11.0.0 || ^12.0.0",
-    "firebase": "^7.0 || ^8.0",
+    "@angular-devkit/architect": "^0.1200.0",
+    "@angular-devkit/core": "^12.0.0",
+    "@angular-devkit/schematics": "^12.0.0",
+    "@angular/common": "^12.0.0",
+    "@angular/compiler": "^12.0.0",
+    "@angular/core": "^12.0.0",
+    "@angular/platform-browser": "^12.0.0",
+    "@angular/platform-browser-dynamic": "^12.0.0",
+    "@angular/router": "^12.0.0",
+    "firebase": "9.0.0-beta.8",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.6.0",
     "firebase-tools": "^8.0.0 || ^9.0.0",
@@ -60,10 +63,11 @@
     "inquirer-autocomplete-prompt": "^1.0.1",
     "jsonc-parser": "^3.0.0",
     "open": "^7.0.3",
-    "rxfire": "^3.9.7 || ^4.0.0",
-    "rxjs": "^6.5.3",
+    "rxfire": "6.0.0-canary.92c6c26",
+    "rxjs": "~6.6.0",
     "semver": "^7.1.3",
-    "tslib": "^1.10.0 || ^2.0.0",
+    "tslib": "^2.1.0",
+    "webpack": "^5.35.0",
     "ws": "^7.2.1",
     "xhr2": "^0.1.4",
     "zone.js": "~0.11.4"
@@ -122,6 +128,6 @@
   },
   "typings": "index.d.ts",
   "resolutions": {
-    "webpack": "5.4.0"
+    "webpack": "^5.35.0"
   }
 }
