@@ -1,31 +1,36 @@
-import { ApplicationRef, Component } from '@angular/core';
-import { FirebaseApp } from '@angular/fire';
-import { debounceTime } from 'rxjs/operators';
+import { ApplicationRef, Component, NgZone } from '@angular/core';
+import { FirebaseApp, FirebaseApps } from '@angular/fire/app';
+import { Auth, AuthInstances, authState } from '@angular/fire/auth';
+import { Firestore, FirestoreInstances, getDoc, doc, DocumentSnapshot } from '@angular/fire/firestore/lite';
+import { DocumentData } from 'rxfire/firestore/lite/interfaces';
+import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
 
 @Component({
   selector: 'app-root',
   template: `
-    <h1>AngularFire kitchen sink</h1>
-    <h2>Primary outlet</h2>
-    <nav>
-      <a [routerLink]="[{ outlets: { primary: [] }}]">Home</a> |
-      <a [routerLink]="[{ outlets: { primary: ['protected'] }}]">Protected</a> |
-      <a [routerLink]="[{ outlets: { primary: ['lazy'] }}]">Lazy</a> |
-      <a [routerLink]="[{ outlets: { primary: ['protected-lazy'] }}]">Protected Lazy</a> |
-      <a [routerLink]="[{ outlets: { primary: ['protected-lazy', 'asdf'] }}]">Protected Lazy Deep</a> |
-      <a [routerLink]="[{ outlets: { primary: ['protected-lazy', '1', 'bob'] }}]">Protected Lazy Deep</a></nav>
+    <pre>{{ (myDocData | async)?.data() | json }}</pre>
     <router-outlet></router-outlet>
-    <h2>Secondary outlet</h2>
-    <nav><a [routerLink]="[{ outlets: { secondary: [] }}]">Home</a> | <a [routerLink]="[{ outlets: { secondary: ['protected'] }}]">Protected</a> | <a [routerLink]="[{ outlets: { secondary: ['protected-lazy'] }}]">Protected Lazy (no anonymous)</a></nav>
-    <router-outlet name="secondary"></router-outlet>
-    <h2>Yet anther outlet</h2>
-    <nav><a [routerLink]="[{ outlets: { tertiary: [] }}]">Home</a> | <a [routerLink]="[{ outlets: { tertiary: ['protected'] }}]">Protected</a></nav>
-    <router-outlet name="tertiary"></router-outlet>
   `,
-  styles: [``]
+  styles: []
 })
 export class AppComponent {
-  constructor(public readonly firebaseApp: FirebaseApp, appRef: ApplicationRef) {
-    appRef.isStable.pipe(debounceTime(200)).subscribe(it => console.log('isStable', it));
+  myDocData: Promise<DocumentSnapshot<DocumentData>>;
+  title = 'sample';
+  constructor(
+    public app: FirebaseApp, // default Firebase App
+    public auth: Auth, // default Firbase Auth
+    public apps: FirebaseApps, // all initialized App instances
+    public authInstances: AuthInstances, // all initialized Auth instances
+    public firestore: Firestore,
+    public firestoreInstances: FirestoreInstances,
+    appRef: ApplicationRef,
+    zone: NgZone,
+  ) {
+    console.log({app, auth, apps, authInstances, firestore, firestoreInstances });
+    authState(auth).subscribe(it => console.log('authState', it));
+    appRef.isStable.pipe(distinctUntilChanged()).subscribe(it => console.log('isStable', it));
+    this.myDocData = getDoc(doc(firestore, 'animals/NJdGQCv1P92SWsp4nSE7'));
+    // firestoreInstance$.subscribe(it => console.log('$', it));
+    // initializeFirestore$.subscribe(it => console.log('init', it));
   }
 }
