@@ -18,13 +18,10 @@
  * BROWSER POLYFILLS
  */
 
-/** IE10 and IE11 requires the following for NgClass support on SVG elements */
+/**
+ * IE11 requires the following for NgClass support on SVG elements
+ */
 // import 'classlist.js';  // Run `npm install --save classlist.js`.
-
-import 'proxy-polyfill/proxy.min.js';
-import 'core-js/stable';
-import 'whatwg-fetch';
-import 'first-input-delay';
 
 /**
  * Web Animations `@angular/platform-browser/animations`
