@@ -1,5 +1,4 @@
-
-import 'firebase/auth'; // removed in build process when not UMD
-
-export * from './auth';
-export * from './auth.module';
+export { Auth, AuthInstances, authInstance$ } from './auth';
+export { provideAuth, AuthModule } from './auth.module';
+export * from './rxfire';
+export * from './firebase';
