@@ -1,7 +1,64 @@
-import { NgModule } from '@angular/core';
-import { AngularFireDatabase } from './database';
+import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders } from '@angular/core';
+import { Database as FirebaseDatabase } from 'firebase/database';
+
+import { AuthInstances } from '@angular/fire/auth';
+import { ɵgetDefaultInstanceOf, ɵmemoizeInstance, ɵAngularFireSchedulers } from '@angular/fire';
+import { Database, DatabaseInstances, DATABASE_PROVIDER_NAME } from './database';
+import { FirebaseApps } from '@angular/fire/app';
+
+export const PROVIDED_DATABASE_INSTANCES = new InjectionToken<Database[]>('angularfire2.database-instances');
+
+export function defaultDatabaseInstanceFactory(_: Database[]) {
+  const defaultDatabase = ɵgetDefaultInstanceOf<FirebaseDatabase>(DATABASE_PROVIDER_NAME);
+  return new Database(defaultDatabase);
+}
+
+export function databaseInstanceFactory(fn: () => FirebaseDatabase) {
+  return (zone: NgZone) => {
+    const database = ɵmemoizeInstance<FirebaseDatabase>(fn, zone);
+    return new Database(database);
+  };
+}
+
+const DATABASE_INSTANCES_PROVIDER = {
+  provide: DatabaseInstances,
+  deps: [
+    [new Optional(), PROVIDED_DATABASE_INSTANCES ],
+  ]
+};
+
+const DEFAULT_DATABASE_INSTANCE_PROVIDER = {
+  provide: Database,
+  useFactory: defaultDatabaseInstanceFactory,
+  deps: [
+    NgZone,
+    [new Optional(), PROVIDED_DATABASE_INSTANCES ],
+  ]
+};
 
 @NgModule({
-  providers: [ AngularFireDatabase ]
+  providers: [
+    DEFAULT_DATABASE_INSTANCE_PROVIDER,
+    DATABASE_INSTANCES_PROVIDER,
+  ]
 })
-export class AngularFireDatabaseModule { }
+export class DatabaseModule {
+}
+
+export function provideDatabase(fn: () => FirebaseDatabase): ModuleWithProviders<DatabaseModule> {
+  return {
+    ngModule: DatabaseModule,
+    providers: [{
+      provide: PROVIDED_DATABASE_INSTANCES,
+      useFactory: databaseInstanceFactory(fn),
+      multi: true,
+      deps: [
+        NgZone,
+        ɵAngularFireSchedulers,
+        FirebaseApps,
+        // Database+Auth work better if Auth is loaded first
+        [new Optional(), AuthInstances ],
+      ]
+    }]
+  };
+}
