@@ -1,8 +1,4 @@
-export * from './database';
-export * from './list/changes';
-export * from './list/create-reference';
-export * from './list/snapshot-changes';
-export * from './list/state-changes';
-export * from './list/audit-trail';
-export * from './observable/fromRef';
-export * from './database.module';
+export { Database, DatabaseInstances, databaseInstance$ } from './database';
+export { provideDatabase, DatabaseModule } from './database.module';
+export * from './rxfire';
+export * from './firebase';
