@@ -1,8 +1,4 @@
-export * from './firestore';
-export * from './firestore.module';
-export * from './collection/collection';
-export * from './collection-group/collection-group';
-export * from './document/document';
-export * from './collection/changes';
-export * from './observable/fromRef';
-export * from './interfaces';
+export { Firestore, FirestoreInstances, firestoreInstance$ } from './firestore';
+export { provideFirestore, FirestoreModule } from './firestore.module';
+export * from './rxfire';
+export * from './firebase';
