@@ -1,5 +1,4 @@
-
-import 'firebase/functions'; // removed in build process when not UMD
-
-export * from './functions';
-export * from './functions.module';
+export { provideFunctions, FunctionsModule } from './functions.module';
+export { Functions, FunctionsInstances, functionInstance$ } from './functions';
+export * from './rxfire';
+export * from './firebase';
