@@ -1,2 +1,3 @@
-export * from './messaging';
-export * from './messaging.module';
+export { MessagingInstances, Messaging, messagingInstance$ } from './messaging';
+export { provideMessaging, MessagingModule } from './messaging.module';
+export * from './firebase';
