@@ -1,2 +1,4 @@
-export * from './remote-config';
-export * from './remote-config.module';
+export { RemoteConfigInstances, RemoteConfig, remoteConfigInstance$ } from './remote-config';
+export { provideRemoteConfig, RemoteConfigModule } from './remote-config.module';
+export * from './rxfire';
+export * from './firebase';
